Refactor sort button rendering and add low to high sort option
Sort button renders with product grid. When low to high button is clicked, the products rerender in lowest to highest price order
